+++ code/cell.modules/pkg.sys/ui.dev/src/types/Host/Host.ts
@@ -1,4 +1,4 @@
-import { t } from './common';
+import { t } from '../common';
 
 /**
  * Display properties of a <Host>.
