+++ code/cell.modules/pkg.sys/ui.dev/src/types/index.ts
@@ -1,7 +1,5 @@
 export * from './types';
 export * from './types.event';
 
-export * from './types.Host';
-export * from './types.Hosted';
-
+export * from './Host';
 export * from './Actions';
